server/requirement.txt
asgiref==3.8.1
Django==5.1.4
PyYAML==6.0.2
sqlparse==0.5.2
tzdata==2024.2
